feat(parquet): upgrade versions and improve code (#42)
* doc(readme): how to change log level
* feat(cargo): upgrade versions
* feat(cargo): upgrade versions for toml & bucket
* feat(example): update tracing
* fix(main): enable opentelemetry if apm feature declared
* fix(makefile): set number of // jobs
* feat(release): speedup the CI

+++ src/connector/mongodb.rs
@@ -522,7 +522,7 @@ impl Metadata {
         let count = connector.clone().len().await?;
 
         info!(count = count, "The counter count with success");
-        Ok(Some(count as usize))
+        Ok(Some(count))
     }
 }
