+++ src/connector/psql.rs
@@ -669,7 +669,7 @@ impl Scan {
         let count = connector.clone().len().await?;
 
         info!(count = count, "The counter count with success");
-        Ok(Some(count as usize))
+        Ok(Some(count))
     }
 }
